pyproject.toml
[tool.poetry]
name = "stoli-bot"
version = "0.1.0"
description = ""
authors = ["1apostoli <79337131+1apostoli@users.noreply.github.com>"]
readme = "README.md"
include = ["stoli_bot"]

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
discord-py-interactions = "^5.11.0"
httpx = "^0.26.0"
notion-client = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
